Update user-manager.py [added logic to prevent user from deleting himself]

+++ source/user_manager.py
@@ -65,7 +65,9 @@ def get_user_username(current_user: Annotated[str, Depends(getCurrentUser)], use
     return user_return
 
 @router.delete("/user/{id}", status_code=status.HTTP_204_NO_CONTENT) #TODO: prevent user from deleteing himself
-def delete_user(current_user: Annotated[str, Depends(getCurrentUser)], id: int, session: SessionDep):
+def delete_user(current_user: Annotated[UserBase, Depends(getCurrentUser)], id: int, session: SessionDep):
+    if current_user.id == id:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="cannot delete current user")
     try:
         user = session.get(User, id) #note: session.get is used to get an obj by primary key only and will return only one row
         session.delete(user)
